create a new component Order and add the contents
