+++ src/components/Restaurant/Restaurant.jsx
@@ -1,19 +1,12 @@
-import { useState } from "react";
-import RestaurantButton from "../RestaurantButton/RestaurantButton.jsx";
+import Order from "@/components/Order/Order.jsx";
 
 const Restaurant = () => {
-  const [orders, setOrders] = useState(0);
-
-  function orderOne() {
-    setOrders(orders + 1);
-  }
+  
   return (
     <section className="restaurant">
       <h3 className="restaurant__heading">Restaurant Orders</h3>
       <ul className="restaurant__list">
-        <li className="restaurant__item">
-          Pizzas: {orders} <RestaurantButton orderOne={orderOne} />
-        </li>
+        <Order />
       </ul>
     </section>
   );
